Added an example app to the package
This example app contains to tabs, one for fruits and one for meals.
You can tap on each fruit/meal and navigate to a detailed view. Once you are in a detail view, simply tap the current tab and you will navigate to the root view. If you are already on the root view and you tap again, the List will scroll to the first item, if possible.